cephadm: move pathify & get_file_timestamp to file_utils
